refactor : SelectionOption 패키지 변경
