+++ src/main/java/view/SelectionOption.java
@@ -1,4 +1,4 @@
-package domain;
+package view;
 
 import java.util.Arrays;
 
@@ -20,7 +20,8 @@ public enum SelectionOption {
                 .filter(select -> select.option.equals(input))
                 .findAny()
                 .orElseThrow(() -> new IllegalArgumentException(
-                        String.format("[ERROR] 잘못된 입력입니다. 입력 가능한 옵션을 입력해 주세요.  입력 가능한 옵션: 1(출석 추가), 2(출석 수정), 3(크루별 출석 기록 확인), 4(제적 위험자 조회), Q(종료)")
+                        String.format(
+                                "[ERROR] 잘못된 입력입니다. 입력 가능한 옵션을 입력해 주세요.  입력 가능한 옵션: 1(출석 추가), 2(출석 수정), 3(크루별 출석 기록 확인), 4(제적 위험자 조회), Q(종료)")
                 ));
     }
